Split inner_product_adj_grad_op_test from inner_product_op_test
